[Base] Expose ParameterGrp type to python
